Added support for Color destructuring syntax #284
