+++ graphics/src/main/kotlin/ktx/graphics/graphics.kt
@@ -3,7 +3,6 @@ package ktx.graphics
 import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.files.FileHandle
 import com.badlogic.gdx.graphics.Camera
-import com.badlogic.gdx.graphics.Color
 import com.badlogic.gdx.graphics.Pixmap
 import com.badlogic.gdx.graphics.PixmapIO
 import com.badlogic.gdx.graphics.g2d.Batch
@@ -15,26 +14,6 @@ import com.badlogic.gdx.utils.ScreenUtils
 import kotlin.contracts.ExperimentalContracts
 import kotlin.contracts.InvocationKind
 import kotlin.contracts.contract
-
-/**
- * Factory methods for LibGDX [Color] class. Allows to use named parameters.
- * @param red red color value.
- * @param green green color value.
- * @param blue blue color value.
- * @param alpha color alpha value. Optional, defaults to 1f (non-transparent).
- * @return a new [Color] instance.
- */
-fun color(red: Float, green: Float, blue: Float, alpha: Float = 1f) = Color(red, green, blue, alpha)
-
-/**
- * Allows to copy this [Color] instance, optionally changing chosen properties.
- * @param red red color value. If null, will be copied from [Color.r]. Defaults to null.
- * @param green green color value. If null, will be copied from [Color.g]. Defaults to null.
- * @param blue blue color value. If null, will be copied from [Color.b]. Defaults to null.
- * @param alpha color alpha value. If null, will be copied from [Color.a]. Defaults to null.
- * @return a new [Color] instance with values copied from this color and optionally overridden by the parameters. */
-fun Color.copy(red: Float? = null, green: Float? = null, blue: Float? = null, alpha: Float? = null) =
-    Color(red ?: r, green ?: g, blue ?: b, alpha ?: a)
 
 /**
  * Automatically calls [Batch.begin] and [Batch.end].
