+++ graphics/src/test/kotlin/ktx/graphics/GraphicsTest.kt
@@ -3,7 +3,6 @@ package ktx.graphics
 import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.backends.lwjgl.LwjglNativesLoader
 import com.badlogic.gdx.files.FileHandle
-import com.badlogic.gdx.graphics.Color
 import com.badlogic.gdx.graphics.OrthographicCamera
 import com.badlogic.gdx.graphics.g2d.Batch
 import com.badlogic.gdx.graphics.glutils.FrameBuffer
@@ -17,8 +16,6 @@ import com.nhaarman.mockitokotlin2.spy
 import com.nhaarman.mockitokotlin2.verify
 import java.io.File
 import org.junit.Assert.assertEquals
-import org.junit.Assert.assertNotEquals
-import org.junit.Assert.assertNotSame
 import org.junit.Assert.assertSame
 import org.junit.Test
 
@@ -26,54 +23,6 @@ import org.junit.Test
  * Tests general utilities related to LibGDX graphics API.
  */
 class GraphicsTest {
-  @Test
-  fun `should construct Color`() {
-    val color = color(red = 0.25f, green = 0.5f, blue = 0.75f, alpha = 0.6f)
-
-    assertEquals(0.25f, color.r)
-    assertEquals(0.5f, color.g)
-    assertEquals(0.75f, color.b)
-    assertEquals(0.6f, color.a)
-  }
-
-  @Test
-  fun `should construct Color with optional alpha`() {
-    val color = color(red = 0.25f, green = 0.5f, blue = 0.75f)
-
-    assertEquals(0.25f, color.r)
-    assertEquals(0.5f, color.g)
-    assertEquals(0.75f, color.b)
-    assertEquals(1f, color.a)
-  }
-
-  @Test
-  fun `should copy Color values`() {
-    val color = Color(0.4f, 0.5f, 0.6f, 0.7f)
-
-    val copy = color.copy()
-
-    assertNotSame(color, copy)
-    assertEquals(0.4f, copy.r)
-    assertEquals(0.5f, copy.g)
-    assertEquals(0.6f, copy.b)
-    assertEquals(0.7f, copy.a)
-    assertEquals(color, copy)
-  }
-
-  @Test
-  fun `should override chosen Color values when copying`() {
-    val color = Color(0.4f, 0.5f, 0.6f, 0.7f)
-
-    val copy = color.copy(red = 0.25f, green = 0.35f, blue = 0.45f, alpha = 0.55f)
-
-    assertNotSame(color, copy)
-    assertEquals(0.25f, copy.r)
-    assertEquals(0.35f, copy.g)
-    assertEquals(0.45f, copy.b)
-    assertEquals(0.55f, copy.a)
-    assertNotEquals(color, copy)
-  }
-
   @Test
   fun `should begin and end Batch`() {
     val batch = mock<Batch>()
